+++ app/Services/CancelFailure.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Services;
+
+enum CancelFailure
+{
+    case DeadlineExceeded;
+    case NotCancellable;
+}